Replaced regex Safari test with multiple "strings.Contains", result:

- reduced allocations on Safari browsers from 9-10/op to 4/op
- 500 to 300 bytes/op
- execution time from ~14000 to ~5000ns/op on my macbookpro i7 2.6