@ryanslade 2016-08-12
Review complete

Would be nice when doing further performance improvements to keep the old and new results and parse them with benchcmp (https://github.com/golang/tools/tree/master/cmd/benchcmp)

The output from that tool can then go in the commit message showing how the commit affected performance.