@jacobpierce 2016-08-12
I'll commit it next time -

```
OLD
BenchmarkAgentSurfer-4	  100000	     15765 ns/op	     321 B/op	       4 allocs/op
BenchmarkEvalSystem-4 	  200000	      5215 ns/op	       0 B/op	       0 allocs/op

NEW
BenchmarkAgentSurfer-4       	  200000	     10859 ns/op	     264 B/op	       3 allocs/op
BenchmarkEvalSystem-4        	  300000	      5090 ns/op	       0 B/op	       0 allocs/op

NEW TESTS (UNCOMPARED)
BenchmarkEvalBrowserName-4   	 1000000	      1657 ns/op	       0 B/op	       0 allocs/op
BenchmarkEvalBrowserVersion-4	  500000	      2480 ns/op	      15 B/op	       0 allocs/op
BenchmarkEvalDevice-4        	 1000000	      1754 ns/op	       0 B/op	       0 allocs/op
BenchmarkParseChromeMac-4    	  200000	      6947 ns/op	     288 B/op	       4 allocs/op
BenchmarkParseChromeWin-4    	  200000	      6654 ns/op	     240 B/op	       3 allocs/op
BenchmarkParseChromeAndroid-4	  100000	     13395 ns/op	     320 B/op	       4 allocs/op
BenchmarkParseSafariMac-4    	  200000	      8384 ns/op	     288 B/op	       4 allocs/op
BenchmarkParseSafariiPad-4   	  200000	      9512 ns/op	     288 B/op	       4 allocs/op